feat: Move lru_cache.py to async_components/cache/
